% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{camundongos}
\alias{camundongos}
\title{Dados experimentais utilizando camundongos}
\format{
Data frame com 93 linhas e 4 variáveis:
\itemize{
\item grupo: 1 - controle; 2 - fungo na dosagem 10mg, 3 - fungo na dosagem 1mg  e 4 - fungo na dosagem 0.1mg
\item peso: peso do rato no início do estudo
\item tempo: tempo de vida (em dias) após a aplicação do tratamento
\item evento: indicadora da ocorrência do evento de interesse (retração)
}
}
\description{
Um estudo laboratorial foi realizado para investigar o efeito protedor do fungo Saccharomycs boulardii em 93 ratos debilitados imunologicamente. Inicialmente, os sistemas imunológicos dos ratos foram debilitados quimicamente e, a seguir, 4 tratamentos (controle e fungo nas dosagens 10mg, 1mg  e 0.1mg) foram alocados aleatoriamente a cada animal. A variável resposta foi o tempo de vida (em dias) após a aplicação do tratamento. O estudo teve por objetivo comparar os tratamentos controlando pelo peso inicial dos ratos.
}
\references{
\insertRef{bookEnrico}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
