% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{compmec}
\alias{compmec}
\title{Tempo de vida de componentes mecânicos}
\format{
Data frame com 40 linhas e 3 variáveis:
\itemize{
\item temperatura: temperatura medida em Kelvin
\item tempo: tempo de vida dos componentes mecânicos (em horas)
\item evento: indicadora da ocorrência do evento de interesse (retração)
}
}
\description{
Um componente mecânico possui a variável temperatura como fator de aceleração para a ocorrência de falha. Com o objetivo de estudar o tempo de vida (em horas) dos componentes, um engenheiro realizou um experimento envolvendo 40 componentes, divididos igualmente em 4 diferentes temperaturas medidas em Kelvin: 300, 350, 400 e 500. A temperatura de 300 graus Kelvin corresponde a temperatura normal de operação do componente, e o modelo de Arrenhius foi utilizado para estudar a confiabilidade dos componentes mecânicos.
}
\references{
\insertRef{bookMarta}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
