% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{diarreia}
\alias{diarreia}
\title{Casos de diarréia em crianças}
\format{
Data frame com 5592 linhas e 11 variáveis:
\itemize{
\item numcri: identificador da criança
\item grupo: vit = receberam vitamina A, pla = placebo
\item sexo: fem = feminino, masc = masculino
\item idade: idade em meses no início do estudo
\item ini: data no início do período sem diarreia
\item fim: data no fim do período sem diarreia
\item diasant: duração em dias do episódio de diarreia anterior
\item mediadej: média de dejeções líquidas ou semilíquidas do episódio de diarreia anterior
\item enum: numeração das observações para cada criança
\item status: 0 = censura, 1 = evento (diarreia)
\item tempo: dias sem diarreia (fim-ini)
}
}
\source{
\url{http://sobrevida.fiocruz.br/diarreia.html}

\insertRef{bookFiocruz}{sobrevivencia}
}
\description{
Estudo sobre o efeito da suplementação de vitamina A na diarreia no nordeste do Brasil em uma coorte de crianças de 6 a 48 meses são referentes ao ensaio comunitário randomizado placebo-controlado, realizado entre dezembro/1990 e dezembro/1991. Banco de dados disponível em \insertRef{bookFiocruz}{sobrevivencia} e cedido pelo Instituto de Saúde Coletiva da Universidade Federal da Bahia. O banco de dados consiste em 860 crianças (426 no grupo placebo e 434 no grupo vitamina), gerando um total de 5.592 registros e 11 variáveis. Esse banco de dados é caracterizado pela presença de múltiplos eventos (cada criança pode ter um ou mais episódios de diarreia ao longo de sua permanência no estudo).
}
\details{
No arquivo, cada linha representa um tempo entre episódios de diarreia. As covariáveis constantes (grupo, sexo e idade) têm seus valores repetidos em todas as linhas referentes a essa criança. As covariáveis que mudam de episódio para episódio são diasant, mediadej e enum.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
