% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{e1684}
\alias{e1684}
\title{E1684 Melanoma Clinical Trial.}
\format{
A data frame with 286 rows and 10 variables:
\itemize{
\item trt: treatment (1 - observation, 2 - IFN)
\item failtime: time to relapse from randomization
\item failcens: failure indicator (1 - failure; 0 - otherwise)
\item survtime: time from relapse to death
\item survcens: failure indicator (1 - failure; 0 - otherwise)
\item age: patient's age (in years)
\item node: 4 nodal categories
\item sex: patient's (0=male, 1=female)
\item ps: patient's performance status (0 - fully active; 1 - other)
\item breslow: Breslow depth (in mm)
}
}
\description{
Two-arm phase III clinical trial conducted by the Eastern Cooperative Oncology Group (ECOG) to comparing high-dose interferon (IFN) to observation (OBS). There were a total of 286 patients diagnosed with melanoma enrolled in the study, which accrued from 1984 to 1990, and the study was unblinded in 1993.
}
\references{
\insertRef{bookIbrahim}{sobrevivencia}
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
