% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggresiduals.R
\name{ggresiduals.coxph}
\alias{ggresiduals.coxph}
\title{ggresiduals method for coxph models}
\usage{
\method{ggresiduals}{coxph}(object, type = c("coxsnell", "martingale", "deviance"), ...)
}
\arguments{
\item{object}{a fitted model object of the class coxph.}

\item{type}{type of residuals used in the plot: coxsnell (default), martingale and deviance.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the desired residual plot.
}
\description{
ggresiduals method for coxph models
}
\details{
This function produces residuals plots of Cox-Snell residuals, martingale residuals and deviance residuals.
}
