% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{hepatite}
\alias{hepatite}
\title{Dados de hepatite}
\format{
Data frame com 29 linhas e 3 variáveis:
\itemize{
\item tempos: tempo até a morte do paciente (em semanas) em decorrência da hepatite ou a perda do acompanhamento.
\item cens: indicadora da ocorrência do evento de interesse
\item grupos: 1 - controle; 2 - esteróide
}
}
\description{
Estudo clínico aleatorizado realizado para investigar da terapia com esteróide no tratameno de hepatite viral aguda. Os 29 pacientes com a doença foram aleatorizados para receber um placebo ou o tratamento com esteróide. Cada paciente foi acompanhado por 16 semanas, ou até a morte, ou até a perda de acompanhamento.
}
\references{
\insertRef{bookEnrico}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
