% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{inca}
\alias{inca}
\title{Dados de pacientes com câncer do INCA}
\format{
Data frame com 862 linhas e 9 variáveis:
\itemize{
\item id: identificador do paciente
\item tempo: tempo de acompanhamento (em dias)
\item status: indicadora do evento (1 = óbito; 0 = censura)
\item sexo: sexo do paciente (Fem = feminino; Male = masculino)
\item idade: idade do paciente (em anos)
\item gptumor: grupo do tumor (Loco = locorregional; Mtx = metastático; Hemato = hematológico)
\item desnut: indicadora de desnutrição (y = sim; n = não)
\item comorbi: indicadora de comorbidade (y = sim; n = não)
\item leucopenia: indicadora de leucopenia (y = sim; n = não)
}
}
\description{
Dados de sobrevivência de 862 pacientes com câncer provenientes do
Instituto Nacional de Câncer (INCA). O banco de dados inclui informações
clínicas e demográficas dos pacientes, bem como o tempo de acompanhamento
e o indicador de ocorrência do evento de interesse (óbito).
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
