% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnbull.R
\name{Inter}
\alias{Inter}
\title{Interval function}
\usage{
Inter(lower, upper)
}
\arguments{
\item{lower}{lower bound of the interval}

\item{upper}{upper bound of the interval}
}
\description{
Interval function
}
