% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{isoladores}
\alias{isoladores}
\title{Tempo de vida de isoladores elétricos}
\format{
Data frame com 60 linhas e 2 variáveis:
\itemize{
\item tempo: tempo de vida dos isoladores (em horas)
\item evento: indicadora da ocorrência do evento de interesse (retração)
}
}
\description{
Um fabricante de em tipo de isolador elétrico quer conhecer o comportamento do seu produto funcionando na temperatura de 200º C. Um teste de vida foi realizado nestas condições usando 60 isoladores elétricos. O teste terminou quando 45 isoladores falharam (censura à direita do tipo II). As 15 unidades que não falharam até o final do teste foram censuradas no tempo t = 2729 horas.
}
\references{
\insertRef{bookMarta}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
