% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{lasers}
\alias{lasers}
\title{Degradação de luz emitida por lasers}
\format{
A data frame with 255 rows and e 3 variables:
\itemize{
\item unidade: indicadora do laser sendo testado
\item hora: hora de medição da corrente funcional dos lasers
\item degradacao: porcentagem no aumento da corrente funcional (relativo à corrente original)
}
}
\description{
Em uma corrente de funcionamento estável, a luz emitida por lasers degrada ao longo do tempo. Para manter a intensidade da luz emitida constante é necessário que a corrente funcional aumente ao longo do tempo. A degradação neste caso é medida através da porcentagem no aumento da corrente funcional (relativo à corrente original). No estudo em questão 15 lasers foram observados no intervalo 250-4000 horas, coletando os valores da corrente funcional a cada 250 horas. Considera-se que o laser irá falhar depois que ele perder 10\% da sua intensidade.
}
\references{
\insertRef{2005Hamada}{sobrevivencia}
}
\author{
Fabio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
