% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{leukUK}
\alias{leukUK}
\title{The Leukemia Survival Data in North-West England}
\format{
A sf/data frame with 1043 rows and 8 variables:
\itemize{
\item time: survival time in days
\item cens: right censoring status 0=censored, 1=dead
\item age: age in years
\item sex:male=1 female=0
\item wbc:white blood cell count at diagnosis, truncated at 500
\item tpi: the Townsend score for which higher values indicates less affluent areas
\item district:administrative district of residence
\item geometry: georeferenced data needed to plot maps.
}

\insertRef{2002Silvia}{sobrevivencia}

A sf/data frame with 1043 rows and 8 variables:
\itemize{
\item time: survival time in days
\item cens: right censoring status 0=censored, 1=dead
\item age: age in years
\item sex:male=1 female=0
\item wbc:white blood cell count at diagnosis, truncated at 500
\item tpi: the Townsend score for which higher values indicates less affluent areas
\item district:administrative district of residence
\item geometry: georeferenced data needed to plot maps.
}

\insertRef{2002Silvia}{sobrevivencia}
}
\description{
A dataset on the survival of acute myeloid leukemia in 1,043 pateietns, first analyzed by \insertCite{2002Silvia}{sobrevivencia}. It is of interest to investigate possible spatial variation in survival after accounting for known subject-specific prognostic factors, which include age, sex, white blood cell count (wbc) at diagnosis, and the Townsend score (tpi) for which higher values indicates less affluent areas. Both exact residential locations of all patients and their administrative districts (24 districts that make up the whole region) are available.

A dataset on the survival of acute myeloid leukemia in 1,043 pateietns, first analyzed by \insertCite{2002Silvia}{sobrevivencia}. It is of interest to investigate possible spatial variation in survival after accounting for known subject-specific prognostic factors, which include age, sex, white blood cell count (wbc) at diagnosis, and the Townsend score (tpi) for which higher values indicates less affluent areas. Both exact residential locations of all patients and their administrative districts (24 districts that make up the whole region) are available.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
