% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{malaria}
\alias{malaria}
\title{Eficácia da imunização pela malária}
\format{
Data frame com 44 linhas e 3 variáveis:
\itemize{
\item tempo: tempo (em dias) decorrido deste a infecção pela malária até a morte do camundongo
\item cens: indicador de falha ou censura (1 - falha; 0 - censura)
\item grupo: 1 - imunizados 30 dias antes da infecção; 2 - sem imunização e com infecção pela malária; 3 - sem imunização e com infecção pela malária e esquistossomose;
\item evento: indicadora da ocorrência do evento de interesse (retração)
}
}
\description{
Estudo experimental realizado com camundongos para verificar a eficácia da imunização pela malária. O estudo, que foi conduzido no Centro de Pesquisas Renee Rachou, Fiocruz-MG, contou com 44 camundongos que foram infectados pela malária. Os camundongos foram aleatoriamente divididos em 3 grupos: grupo: 1 - imunizados 30 dias antes da infecção; 2 - sem imunização e com infecção pela malária; 3 - sem imunização e com infecção pela malária e esquistossomose.
}
\references{
\insertRef{bookEnrico}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
