% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mama}
\alias{mama}
\title{Câncer de mama}
\format{
Data frame com 94 linhas e 4 variáveis:
\itemize{
\item esquerda: limite inferior do intervalo (em meses)
\item direita: limite superior do intervalo (em meses)
\item terapia: radioterapia ou radioterapia + quimioterapia
\item evento: indicadora da ocorrência do evento de interesse (retração)
}
}
\description{
Estudo retrospectivo realizado com 94 mulhres, das quais 46 receberam o tratamento somente com radioterapia e 48 com radioterapia e quimioterapia. O evento de interesse foi definido como a ocorrência da primeira (moderada ou severa) retração de uma mama. As pacientes foram inicialmente observadas a cada 4-6 meses e, assim que ficavam melhores o intervalo entre as visitas aumentava.
}
\references{
\insertRef{bookEnrico}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
