% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{nwengland}
\alias{nwengland}
\title{The Leukemia Survival Data in North-West England}
\format{
A sf/data frame with 2 rows and 2 variables:
\itemize{
\item district: district number
\item geometry: georeferenced data needed to plot maps.
}

\insertRef{2002Silvia}{sobrevivencia}
}
\description{
Georeferenced data on 24 administrative districts in the North-West of England used in \insertCite{2002Silvia}{sobrevivencia}.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
