% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{oportunista}
\alias{oportunista}
\title{Ocorrências de doenças oportunistas em pessoas diagnosticadas com AIDS}
\format{
Data frame com 1195 linhas e 11 variáveis:
\itemize{
\item reg: identificador do paciente
\item sex: F = feminino, M = masculino
\item esc: 0 = sem escolaridade, 1 = até quatro anos de estudo, 2 = ensino fundamental, 3 = ensino médio, 4 = ensino superior
\item idade: idade em anos na entrada do estudo
\item udi: uso de drogas injetáveis: 0 = não, 1 = sim
\item sexual: comportamento sexual: 0 = não, 1 = sim
\item ini: data da entrada no estudo (diagnóstico de Aids)
\item oport: doença oportunista: Candida = candidíase, Tuberc = tuberculose, Hemato = algumas alterações hematológicas, Herpes = herpes zoster, Pneumo = pneumonia por Pneumocistis
\item fim: data do diagnóstico do evento relacionado à imunodepressão
\item status: 0 = censura, 1 = ocorrência do evento
\item tempo: fim - ini
}
}
\source{
\url{http://sobrevida.fiocruz.br/aidsoport.html}

\insertRef{bookFiocruz}{sobrevivencia}
}
\description{
Os dados utilizados para o tempo até a ocorrência de doenças oportunistas provêm de uma coorte de pacientes HIV positivos atendidos no Hospital Universitário Gaffrée e Guinle (Unirio). As variáveis foram obtidas a partir dos prontuários clínicos. Registrou-se, para cada paciente com Aids, o tempo até a ocorrência de cada doença ou sintomatologia caracteristicamente relacionada à imunodepressão: candidíase, tuberculose, alguns sinais hematológicos, herpes zoster, pneumonia por Pneumocistis.
}
\details{
Os dados estão armazenados de maneira que cada linha do arquivo indica o tempo do diagnóstico de Aids até a ocorrência de uma doença oportunista.
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
