% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengamma.R
\name{pggamma}
\alias{pggamma}
\title{Generalized gamma distribution}
\usage{
pggamma(q, alpha, gamma, kappa, log = FALSE, lower.tail = TRUE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{alpha}{scale parameter (alpha > 0).}

\item{gamma}{shape parameter (gamma > 0).}

\item{kappa}{shape parameter (kappa > 0).}

\item{log}{logical; if TRUE, probabilities are returned on the log scale. Default is FALSE.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x), otherwise P(X > x).}
}
\value{
A numeric vector of (log) probabilities.
}
\description{
Generalized gamma distribution
}
