% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{pulmao}
\alias{pulmao}
\title{Câncer de pulmão em camundongos}
\format{
Data frame com 144 linhas e 3 variáveis:
\itemize{
\item tempo: tempo até a morte do camundongo (em dias)
\item tumor: presença/ausência do tumor
\item ambiente: ambiente convencional ou ambiente livre de germes
}
}
\description{
Experimento de tumorigenicidade sobre tumores pulmonares para 144 camundongos RFM machos. Os dados consistem no tempo até a morte de cada animal medido em dias, e um indicador de presença ou ausência de tumor pulmonar no momento da morte. O experimento envolve dois tratamentos, ambiente convencional (96 camundongos) e ambiente livre de germes (48 camundongos). Tumores pulmonares em camundongos RFM são predominantemente não letais, o que significa que a ocorrência de um tumor não não alterar a taxa de mortalidade.
}
\references{
\insertRef{2006Sun}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
