% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R, R/ggresiduals.R, R/survfit.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{augment}
\alias{ggresiduals}
\alias{survfit}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom:reexports]{augment}}}

  \item{survival}{\code{\link[survival]{survfit}}}

  \item{survstan}{\code{\link[survstan]{ggresiduals}}}
}}

