% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sinusite}
\alias{sinusite}
\title{Sinusite em pacientes com HIV}
\format{
Data frame com 133 linhas e 12 variáveis:
\itemize{
\item pac: identificador do paciente
\item id: idade do paciente (em anos)
\item sex: sexo do paciente (0 - masculino; 1 feminino)
\item grp: grupo de risco (1 - paciente HIV soronegativo; 2 - paciente HIV soropositivo assintomático; 3 - paciente com ARC; 4 paciente com AIDS)
\item ti: tempo inicial (em dias)
\item tf: tempo final (em dias)
\item cens: indicadora de falha ou censura
\item cd4: contagem de células cd4 medida no início do estudo
\item cd8: contagem de células cd8 medida no início do estudo
\item ats: atividade sexual (1 - homossexual; 2 - bissexual; 3 heterossexual)
\item ud: use de droga injetável (1 - sim; 2 - não)
\item ac: uso de cocaína por aspiração (1 - sim; 2 - não)
}
}
\description{
Estudo realizado no Departamento de Otorrinolaringologia da UFMG, entre março de 1993 e fevereiro de 1995, com o objetivo de verificar a hipótese de que a infecção pelo HIV aumenta o risco de sinusite. Participaram do estudo 91 pacientes infectados com HIV e 21 pacientes não infectados com HIV.
}
\references{
\insertRef{bookEnrico}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
