% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{transformadores}
\alias{transformadores}
\title{Manutenção preventiva de transformadores}
\format{
Data frame com 61 linhas e 5 variáveis:
\itemize{
\item tranformador: indicadora do transformador sendo acompanhado
\item unidade: indicadora do tranformador sendo acompanhado (transformadores que passam por manutenção preventiva são considerados novas unidades)
\item preventiva: variável indicadora (1 - ocorrência de manutenção preventiva; 0 -  caso contrário)
\item tempo: tempo até a falha do transformador (em horas)
\item evento: indicadora da ocorrência do evento de interesse (falha do tranformador)
}

Data frame com 61 linhas e 5 variáveis:
\itemize{
\item tranformador: indicadora do transformador sendo acompanhado
\item unidade: indicadora do tranformador sendo acompanhado (transformadores que passam por manutenção preventiva são considerados novas unidades)
\item preventiva: variável indicadora (1 - ocorrência de manutenção preventiva; 0 -  caso contrário)
\item tempo: tempo até a falha do transformador (em horas)
\item evento: indicadora da ocorrência do evento de interesse (falha do tranformador)
}
}
\description{
Uma companhia de energia elétrica deseja estabelecer uma política de manutenção preventiva para seus transformadores, visando minimizar o custo esperado associado a sua manutenção. Tal política de manutenção deve levar em conta ambos os custos decorrentes da paralização programada (manutenção preventiva) e não programada (falha e posterior reparo do tranformador). Com esse objetivo, 30 transformadores foram colocados sob teste e concluiu-se que o tempo ótimo entre manutenções preventivas é igual a 6285 horas.

Uma companhia de energia elétrica deseja estabelecer uma política de manutenção preventiva para seus transformadores, visando minimizar o custo esperado associado a sua manutenção. Tal política de manutenção deve levar em conta ambos os custos decorrentes da paralização programada (manutenção preventiva) e não programada (falha e posterior reparo do tranformador). Com esse objetivo, 30 transformadores foram colocados sob teste e concluiu-se que o tempo ótimo entre manutenções preventivas é igual a 6285 horas.
}
\references{
\insertRef{2007Gilardoni}{sobrevivencia}

\insertRef{2007Gilardoni}{sobrevivencia}
}
\author{
Fábio N. Demarqui \email{fndemarqui@est.ufmg.br}
}
\keyword{datasets}
