% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnbull.R
\name{turnbull}
\alias{turnbull}
\title{Turnbull estimator for interval-censored survival data}
\usage{
turnbull(formula, data, eps = 0.001, iter.max = 200, verbose = FALSE, ...)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which yppe is called.}

\item{eps}{eps}

\item{iter.max}{maximum number of iterations of the algorithm}

\item{verbose}{default = FALSE}

\item{...}{further arguments passed to other methods}
}
\value{
a data.frame with the estimated survivals
}
\description{
Turnbull estimator for interval-censored survival data
}
\examples{
\donttest{
library(sobrevivencia)
library(tidyverse)

data(mama)
mama <- mama \%>\%
 mutate(
   direita = ifelse(is.na(direita), Inf, direita)
 )


tbe1 <- turnbull(Inter(esquerda, direita)~1, data = mama)
tbe2 <- turnbull(Inter(esquerda, direita)~terapia, data = mama)

ggplot(tbe2, aes(x=time, y=survival, color = terapia)) +
 geom_step()

ggplot(tbe1, aes(x=time, y=survival)) +
 geom_step()
}
}
