% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gengamma.R
\name{rggamma}
\alias{rggamma}
\title{#' @export
qggamma2 <- function(p, nu, gamma, kappa){
q <- qgamma(p, shape = nu, scale = 1)
x <- gamma*(q^(1/kappa))
return(x)
}}
\usage{
rggamma(n, alpha, gamma, kappa)
}
\description{
#' @export
qggamma3 <- function(p, nu, gamma, kappa){
q <- qgamma(p, shape = nu, scale = 1)
x <- gamma*(q^(1/kappa))
return(x)
}
}
